fix: Use health checks to drive service dependencies in Docker compose files.
Adds build context to easily build app:build images.

Disable syncserver in E2E tests. Migrations can be racy since the
tests start their own server. This wasn't a problem with hardcoded
sleep, but there's a higher chance both syncserver and E2E tests find
the DB to be alive at the same time now.

We cannot use the service_completed_successfully condition for
sync-db-setup, because of --abort-on-container-exit in CI.